Add a dangling CNAME check in
When using a CNAME, it's possible point a record at a location and later
have that destination disappear or stop responding as you would expect.
In the past we haven't been able to detect it but this adds a new check
the CNAME types that attempts to resolve the full chain of lookup values
for a given value. The idea is that if you look up a CNAME and get back
a NXDOMAIN exception from the source (through `getaddresses`) then
it's potentially a dangling DNS record and you should be alerted to
it.

Here are two examples of it in action:

- Black hole end result but our CNAME still present

```
$ dig -t cname _ca3-abcd123.well-groomed-alpaca.com @1.1.1.1 +short
dcv.digicert.com.
```

```
$ ruby -rdnsruby -e 'p Dnsruby::Resolv.getaddresses("_ca3-abcd123.well-groomed-alpaca.com")'
...
/Users/jacob/.asdf/installs/ruby/2.6.6/lib/ruby/gems/2.6.0/gems/dnsruby-1.61.3/lib/dnsruby/DNS.rb:298:in `send_query': Dnsruby::NXDomain (Dnsruby::NXDomain)
```

- CNAME that resolves all the way to the end

```
$ dig -t cname _ca3-realcname.well-groomed-alpaca.com @1.1.1.1 +short
example.com.
```

```
$ ruby -rdnsruby -e 'p Dnsruby::Resolv.getaddresses("_ca3-realcname.well-groomed-alpaca.com")'
[#<Dnsruby::IPv4 93.184.216.34>]
```